Handle on axis DDA traversal.
The DDA traversal fails when dividing by 0 which happens when the ray direction is 0 in any axis.
When this happens, we can know that inf > finite so there is no need to test for it.
Previously testing for inf < inf would cause undefined results on axis traces would end in rays terminating early or never terminating